+++ CHANGELOG.md
@@ -61,3 +61,4 @@
 
 ## 0.5.2
 - Add assigned questions to practice report
+- Clean up closing of practice and practice report windows when practice reports are successfully sent