+++ CHANGELOG.md
@@ -62,3 +62,4 @@
 ## 0.5.2
 - Add assigned questions to practice report
 - Clean up closing of practice and practice report windows when practice reports are successfully sent
+- Add progress bar to sending of reports